Adds 'filter-by-series' to audio search
This filter will filter audios based on whether they are attached to a
series or not.
'true' will only return audios that are attached to a series, 'false'
will only return audios that are NOT attached to a series, not
specifying 'filter-by-series' will return audios both with and without
series attachments.