Set `track_total_hits` to true for all elasticsearch searches
We set the `track_total_hits` parameter to true to accurately respond
with total hits of the search query. This only affects queries that are
above 10000 results.

I think the performance can be affected in these queries, but i assume
that happened before the upgrade as well so i think it is fine.

https://www.elastic.co/guide/en/elasticsearch/reference/7.17/search-your-data.html#track-total-hits