image-api: Structure metadata for multiple images
This patch structures `domain.Image` as a part of
`domain.ImageMetaInformation` so that we may have separate images for
different languages for a single image.

This includes a TODO since `ImageRepository.getImageFromFilePath` now
uses a slow jsonb query which should be replaced by restructuring the
model to allow for faster lookups.