Move shared image controller code to base class
Previously `ImageControllerV2` simply inherited from `ImageControllerV3`
This worked mostly fine, but when generating swagger-doc for
V2 we also generated doc for the V3 endpoints in the V2 context.
So V3 was generated one time too many. This workaround avoid that by not
running the V3 endpoints for V2 by using a base class for sharing code.