learningpath-api: Split folders and resources into two separate lists
Makes frontend-handling of stuff easier.
Only drawback i can see is that if/when we want to sort across both
resources and folders we would need to expose a `rank` flag or something
that the frontend would need to sort by manually. Probably not a big
deal if type-handling and such is easier with the new layout :^)