Moved scalatra package content to common package
The more code we move to common package the more we observe that plenty of that code depends on scalatra.
We moved scalatra content to common instead of making common depend on scalatra.
The downside is that every application that depends on common will get scalatra as a dependency.
Luckily most of our applications depend on scalatra from before, only frontpage-api does not.