Stop failing on multiple parent topics when building context
Since this isn't supposed to happen, we previously failed when
encountering this. This does however happen and I can't seem to figure
out exactly why. For now lets include a FIXME for this and accept
multiple topic paths (and log it obviously).

Also move some code around for small readability improvements.