Move every thread-local variable into `RequestInfo`
And only use `RequestInfo` to set thread context.
This gives us a single source of truth for this and makes it harder to
make mistakes with this and spawning `Future` and `IO` in the future.

This includes moving the `scalatra` part of `common` to `network` as
that makes the dependency graph simpler.