Convert `DraftStatus` from standard scala enum to enumeratum
There was a _WEIRD_ bug where json4s was unable to serialize the
`DraftStatus` in some big objects, but not consistently. It could even
work and fail in the same list. Which resulted in json that looked like
this: `["IN_PROGRESS", {}]` which is very unfortunate.

I was unable to figure out how to reproduce it consistently, but it was
fixed by converting the enum.

I think enumeratum is quite a bit better than the default scala enum
either way so it isn't much of a loss, but it is a bit scary that we
don't know what caused it.

In the future I would be happy to remove all usage of json4s and move to
a more non-exception-throwing json serialization library.