draft-api: Fix bug which resulted in 409 when publishing
The problem was that we did not use the same transaction when fetching
the draft after the original draft had been saved, which resulted in the
revision fetched in the `withId` call to update being the old one.

This patch updates `withId` to take `DBSession` parameter and use it
instead of providing its own session.