audio-api: Handle empty-query parameter when searching
Previously when using scalatra passing `?query=` resulted in `None` for
the `query` query parameter.
Tapir handles this differently which results in the `query` param being
`Some("")`. This makes sense, but is unwanted in this case.

This patch introduces `MaybeNonEmptyString` which can be used as a query
parameter with a codec that makes `?query=` result in
`MaybeNonEmptyString(underlying=None)`.