Implement `AutoEnvPlugin`
This patch introduces a custom sbt-plugin: `AutoEnvPlugin` which loads
environment and/or system properties based on files in the
`envFilePaths` settingKey, this defaults to:
	[".env", "$module/.env", ".env.$module"]

This means that one could keep a shared `.env` file for all subprojects
on root and a separate one for each sub-module in their respective
folder and it would all work!

With and without forking enabled (Since we use properties to load
environment variables in the applications)!