common: Improve `NDLADate` json4s error handling / error message
Json4s appearently needs the exception to be a `MappingException` to
generate its regular error messages so lets use that.

This also reverts to the weird behavior when parsing `Option[NDLADate]`
that makes a invalid date parse as `None`. Hopefully we mostly avoid
relying in this functionality because it will most certainly break with
other, more sensible parsers :^)