Add properties for AWS S3 regions
This patch adds a environment variable for the s3 buckets region.
Instead of attempting to find the region with ec2 metadata by default,
we simply default to `eu-west-1`. This makes more sense as there is
really no technical reason why the s3 bucket has to exist in the same
region as the application.

The old behavior is still available by specifying `auto` as the
environment variable, but i believe the default region with no
environment variable specified will be sufficient.