learningpath-api: Add support for string-list config parameters
This also changes the output/input of the api to be actual `boolean` or
`string[]` types rather than string. For those types.
This is technically breaking, but since this is such a small change with
negligible consequences, _and_ it improves the api, lets just do the break and fix it in
graphql-api without versioning.