Make `CommaSeparatedList` to work around empty default values
When using `sttp.tapir.model.CommaSeparated` directly we end up with
some string as default value when using defaults.

This results in swagger thinking there should be one element with an
empty string in the request.
Wrapping it in an option allows us to have default value of nothing, and
the implicit class allows the migration between
`Option[CommaSeparated[...]]` and `CommaSeparated[...]` to be quite easy.

See: https://github.com/softwaremill/tapir/issues/3581