learningpath-api: Stop fetching cover photos from image-api
This would result in some slow api-calls. Especially ones with a big
number of learningpaths in the results.
IE fetching `/learningpath-api/v2/learningpaths/status/UNLISTED` would
take ~25 seconds. We should probably consider implementing pagination
for such cases as well, but this makes the problem much less of a
problem.

Lets make the caller (graphql-api) handle fetching this instead if they
need the image adress to have file extensions or similar.

Others can simply use the constructed url with the id.