search-api: Add possible sort values to schema
This will add `allowableValues` to openapi doc.
In an ideal world the `SearchParams` types would simply type the sort
parameter as `Option[Sort]` and this would work by itself, but since
that messes with the frontend relying on the sort param being the same
type across apis generated typescript-params this will do.