Replace AWS SDK 1.x with 2.28.9
This is a major version that is very different from the previous one so
a bunch of functions has been rewritten.

We now utilize a wrapper class `NdlaS3Client` to leave a bit of the
abstraction up to us.

While we still aren't fully disconnected from the aws-sdk java types we
are a bit more loosely coupled.