audio-api: Hack around openapi generation of ids endpoint
There was a bug with how the output type of `/audio-api/v1/audios/ids/`
were generated.

For some reason it is fixed when using an `Array` instead of a `List` as
the output.

Not sure of the reason, I assume it has to do with the recursive type of
`AudioMetaInformationDTO` + `SeriesDTO`, but couldn't get it to work in
a satisfying way, so lets just hack around it with `Array` for now.