Eagerly load properties, but don't throw before starting server
This patch introduces a abstraction called `Prop` which we use to load
required properties. These will be loaded eagerly so we can crash on
startup if they are missing. But they also allow us to run parts of the
project without them.

Allows us to run applications with `--save-swagger` argument without
variables that isn't required until runtime.