Eagerly load, but lazily fail environment variables on startup (#641)
* ci: check formatting in test compilation target as well

* Make brightcove environment variables lazy

This is so we can run the type generation script without the setup.
Makes it easier to run the type generation locally without setting up a
bunch of needless variables.

This does however make detecting missing environment variables less
likely at startup. We should look into how we can make those detected at
startup, but not when using `--save-swagger` arguments.

* Eagerly load properties, but don't throw before starting server

This patch introduces a abstraction called `Prop` which we use to load
required properties. These will be loaded eagerly so we can crash on
startup if they are missing. But they also allow us to run parts of the
project without them.

Allows us to run applications with `--save-swagger` argument without
variables that isn't required until runtime.

* testbase: adds a testbase subproject

This should allow us to use dependencies from `common` in
`scalatestsuite` while keeping the base suite available for `common`
itself.

* Make `Prop` not generic

Caused more trouble than it was worth with this implementation. Might be
worth it to revisit in the future.

* Make `Prop` testing work with initial and updated values

* Convert `Prop` to AST & add some tests

I would argue slightly more readable and extendable

* Make `Prop` generic

This makes it so props can be generic and return other types than string
:^)

* testing: Split integration suites into different traits

* Fix error message in `propFromTestValue`