refactor(mill): use `SbtModule` instead of changing `moduleDir`
It was recommended to rather use `SbtModule` to allow the sbt folder
structure rather than changing the `moduleDir` of the test module.

I do not think it makes a big difference, but at least we're following
recommended practices.

In the future we can go back to `ScalaModule` by moving the sources to
the basic mill folder structure[1], but lets keep the changes to the
actual sources minimal until we have battle-tested mill a little bit
more.

- [1] https://mill-build.org/mill/scalalib/intro.html#_simple_scala_module