feat: add `dependency-graph` module
This patch introduces the `dependency-graph` module.
This module is a linting module that utilizes scalameta to detect
cyclical dependencies in our internal classes.

Ideally this would probably be a build system plugin, but since there is
a issue with importing scalameta in the mill-build this is a standalone
module for now.

https://github.com/com-lihaoyi/mill/issues/5759