fix: props eager loading, but lazy failing
When refactoring the DI setup the lazily loaded `DataSource` were no
longer lazily loaded, which meant that running `generateTypescript`
failed on CI for all components that required properties on startup.

This patch fixes that by passing the `dataSource` and a few other
dependencies as a by-name parameter wherever previously used on startup.