feat: rethrowing of interrupted exceptions (#831)
* feat: rethrowing of interrupted exceptions

since some libraries we use swallow interrupted states with their
internal exception handling we want to "fix" this by rethrowing the
interrupted exceptions so that they are handeled in our outtermost
exception handler instead of following the regular error flow.

* feat: also deal with interrupted threads in `ErrorHandling`

Ideally this shouldn't happen, but since some not-so-cool libraries
wraps the interrupt exception before throwing it we need to do some
magic handling.

* refactor: attempt to improve flakiness of controller tests